fix(documentation): improve example
