+++ package.json
@@ -27,7 +27,7 @@
     "test": "npm run lint && npm run test:web",
     "test:web": "NODE_ENV=test jest --coverage",
     "test:clean": "rimraf ./coverage",
-    "test:watch": "npm run test -- --watch",
+    "test:watch": "npm run test -- --watch",s
     "lint": "eslint src",
     "prepublish": "npm run build",
     "lint-staged": "lint-staged",
